+++ src/xml/index.ts
@@ -1,2 +1,3 @@
 export * from './build'
+export * from './inspect-xml'
 export * from './parse'